enable Playlists for oauth login flow
Several places use the `backend.auth` flag to check if we're logged in
or not. Since the API responds simlarly whether using the OAuth login or
the old cookie login, this unifies those checks back under the
`backend.auth` flag.

To simplify setting it the flag, and since `_ytmusicapi_oauth_json` was
only used for one function, I replaced it with direct references to the
original `config['ytmusic']['oauth_json']`.

Finally, checking before reading from `config` avoids crashes when only
one or the other of `auth_json` or `oauth_json` is present in the config
file